skal ikke ha interface her.
