+++ sheepfarmer3000/src/main/java/org/prosjekt/helperclasses/Attack.java
@@ -3,22 +3,21 @@ package org.prosjekt.helperclasses;
 import java.io.Serializable;
 
 import org.joda.time.DateTime;
-import org.prosjekt.helperclasses.impl.SheepImpl;
 
 public class Attack implements Serializable {
-	private SheepImpl sheep;
+	private Sheep sheep;
 	private String latitude;
 	private String longitude;
 	private DateTime timestamp;
 	
-	public Attack(SheepImpl sheep, String latitude, String longitude) {
+	public Attack(Sheep sheep, String latitude, String longitude) {
 		this.sheep = sheep;
 		this.latitude = latitude;
 		this.longitude = longitude;
 		this.timestamp = DateTime.now();
 	}
 
-	public SheepImpl getSheep() {
+	public Sheep getSheep() {
 		return sheep;
 	}
