+++ sheepfarmer3000/src/main/java/org/prosjekt/helperclasses/Passhash.java
@@ -4,11 +4,27 @@
  */
 package org.prosjekt.helperclasses;
 
+import java.io.Serializable;
+
+
 /**
  *
  * @author Christoffer <christofferbuvik@gmail.com>
  */
-public interface Passhash {
-    public String getPasshash(int farmerid);
-    public void setPasshash(int farmerid, String passhash);
+public class Passhash implements Serializable{
+     private String passhash;
+    private int farmerid;
+
+    public Passhash(int farmerid) {
+        this.farmerid = farmerid;
+    }
+    
+    public String getPasshash(int farmerid) {
+        return passhash;
+    }
+
+    public void setPasshash(int farmerid, String passhash) {
+        this.passhash = passhash;
+    }
+    
 }
